- Added GlobalExceptionHandler - Added custom exceptions - Document all APIs using OpenAPI - Generate Swagger UI for easy testing.

+++ src/main/java/com/example/taskmanagementsystem/service/task/AdminTaskService.java
@@ -4,10 +4,11 @@ import com.example.taskmanagementsystem.dto.task.TaskRequestDto;
 import com.example.taskmanagementsystem.dto.task.TaskResponseDto;
 import com.example.taskmanagementsystem.entity.Task;
 import com.example.taskmanagementsystem.entity.User;
+import com.example.taskmanagementsystem.exception.MissingQueryParameterException;
+import com.example.taskmanagementsystem.exception.ResourceNotFoundException;
 import com.example.taskmanagementsystem.mapper.TaskMapper;
 import com.example.taskmanagementsystem.repository.TaskRepository;
 import com.example.taskmanagementsystem.repository.UserRepository;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,10 +25,14 @@ public class AdminTaskService {
 
     public TaskResponseDto createTask(TaskRequestDto taskRequest) {
         User author = userRepository.findById(taskRequest.getAuthorId())
-                .orElseThrow(() -> new EntityNotFoundException("Author not found"));
-        User executor = taskRequest.getExecutorId() != null
-                ? userRepository.findById(taskRequest.getExecutorId()).orElse(null)
-                : null;
+                .orElseThrow(() -> new ResourceNotFoundException("Author not found with ID: " + taskRequest.getAuthorId()));
+
+        User executor = null;
+        if (taskRequest.getExecutorId() != null) {
+            executor = userRepository.findById(taskRequest.getExecutorId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Executor not found with ID: " + taskRequest.getExecutorId()));
+        }
+
         Task task = Task.builder()
                 .title(taskRequest.getTitle())
                 .description(taskRequest.getDescription())
@@ -43,10 +48,13 @@ public class AdminTaskService {
 
     public TaskResponseDto updateTask(Long id, TaskRequestDto taskRequest) {
         Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Task not found"));
-        User executor = taskRequest.getExecutorId() != null
-                ? userRepository.findById(taskRequest.getExecutorId()).orElse(null)
-                : null;
+                .orElseThrow(() -> new ResourceNotFoundException("Task not found with ID: " + id));
+
+        User executor = null;
+        if (taskRequest.getExecutorId() != null) {
+            executor = userRepository.findById(taskRequest.getExecutorId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Executor not found with ID: " + taskRequest.getExecutorId()));
+        }
 
         task.setTitle(taskRequest.getTitle());
         task.setDescription(taskRequest.getDescription());
@@ -65,14 +73,18 @@ public class AdminTaskService {
     }
 
     public void deleteTask(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Cannot delete. Task not found with ID: " + id);
+        }
         taskRepository.deleteById(id);
     }
 
     public TaskResponseDto assignTaskExecutor(Long id, Long executorId) {
         Task task = taskRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Task not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Task not found with ID: " + id));
         User executor = userRepository.findById(executorId)
-                .orElseThrow(() -> new EntityNotFoundException("Executor not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Executor not found with ID: " + executorId));
+
         task.setExecutor(executor);
         Task updatedTask = taskRepository.save(task);
         return taskMapper.toTaskResponseDto(updatedTask);
@@ -81,14 +93,18 @@ public class AdminTaskService {
     public List<TaskResponseDto> getFilteredTasks(Long authorId, Long executorId, String status,
                                                   String priority, int page, int size) {
         PageRequest pageable = PageRequest.of(page, size);
+
+        if (authorId == null && executorId == null) {
+            throw new MissingQueryParameterException("Either authorId or executorId must be provided.");
+        }
+
         Page<Task> tasksPage;
         if (authorId != null) {
             tasksPage = taskRepository.findByAuthorIdWithFilters(authorId, status, priority, pageable);
-        } else if (executorId != null) {
-            tasksPage = taskRepository.findByExecutorIdWithFilters(executorId, status, priority, pageable);
         } else {
-            throw new IllegalArgumentException("Either authorId or executorId must be provided");
+            tasksPage = taskRepository.findByExecutorIdWithFilters(executorId, status, priority, pageable);
         }
+
         return tasksPage.stream()
                 .map(taskMapper::toTaskResponseDto)
                 .toList();
